chore: refactor client.ts, extract query building methods onto separate class
